refatored UnionDB to its own module
